refactor: move getSpotifyUserAlbum to util
